Migrate tool_preparation phase + summary helper to actor_loop_flow
Move drive_actor_loop_tool_preparation_phase and its private helper
record_actor_loop_tool_call_summaries out of impl Agent into
actor_loop_flow.rs. The summary helper becomes a module-private free
fn in actor_loop_flow.rs (not pub(super)) since only the phase
function consumes it.

Visibility upgrades on Agent methods:

- pub(super) on prepare_tool_call
- pub(super) on effective_tool_policy

Drop now-unused production imports from turn.rs and demote
FocusedEditBatchAction to test-only (in-file regression tests still
use it; production consumers all moved with the phase function):

- ActorLoopRejectedToolBatchArgs / handle_actor_loop_rejected_tool_batch
  (used only inside moved fn)
- effective_tool_batch_action_with_scope (used only inside moved fn)
- FocusedEditBatchAction (production use moved; #[cfg(test)] import
  retained for in-file tool_policy regression tests)

Part of #680 / #681 incremental method migration.
